moved profile elements into separate files
