+++ birthdaytracker/lib/widgets/bordered_birthday.dart
@@ -1,0 +1,25 @@
+import 'package:birthdaytracker/models/birthdayprofile.dart';
+import 'package:flutter/material.dart';
+
+class BorderedBirthday extends StatelessWidget {
+  final BirthdayProfile profile;
+
+  BorderedBirthday(this.profile, {super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Container(
+          decoration: BoxDecoration(
+              color: Colors.grey[300], borderRadius: BorderRadius.circular(8)),
+          child: Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Text(
+              profile.getBirthdayString(),
+              style: TextStyle(fontSize: 18),
+            ),
+          )),
+    );
+  }
+}
